Add integration tests for uninstallation of the plugin.
